Move client definitions into separate file.
