+++ Mobile Buy SDK/Mobile Buy SDK/Operations/BUYCheckoutOperation.h
@@ -26,14 +26,13 @@
 
 #import <Buy/BUYGroupOperation.h>
 #import <Buy/BUYStatusOperation.h>
+#import <Buy/BUYClientTypes.h>
 NS_ASSUME_NONNULL_BEGIN
 
 @class BUYClient;
 @class BUYCheckout;
 
 @protocol BUYPaymentToken;
-
-typedef void (^BUYCheckoutOperationCompletion)(BUYCheckout * _Nullable checkout, NSError * _Nullable error);
 
 @interface BUYCheckoutOperation : BUYGroupOperation
